+++ packages/grid/src/dashboard-grid/grid-item.tsx
@@ -166,7 +166,7 @@ export function GridItem( {
 
 			<div
 				ref={ setActivatorNodeRef }
-				{ ...attributes }
+				{ ...( dragDisabled ? {} : attributes ) }
 				{ ...listeners }
 				style={ {
 					height: '100%',